+++ pinot-query-planner/src/main/java/org/apache/pinot/query/parser/CalciteExpressionParser.java
@@ -18,14 +18,12 @@
  */
 package org.apache.pinot.query.parser;
 
-import com.google.common.base.Preconditions;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashSet;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Set;
-import org.apache.calcite.rel.type.RelDataType;
 import org.apache.calcite.sql.SqlBasicCall;
 import org.apache.calcite.sql.SqlDataTypeSpec;
 import org.apache.calcite.sql.SqlIdentifier;
@@ -36,13 +34,10 @@ import org.apache.calcite.sql.SqlNodeList;
 import org.apache.calcite.sql.fun.SqlCase;
 import org.apache.calcite.sql.parser.SqlParseException;
 import org.apache.calcite.sql.parser.SqlParser;
-import org.apache.commons.lang3.StringUtils;
 import org.apache.pinot.common.request.Expression;
 import org.apache.pinot.common.request.ExpressionType;
 import org.apache.pinot.common.request.Function;
-import org.apache.pinot.common.request.PinotQuery;
 import org.apache.pinot.common.utils.request.RequestUtils;
-import org.apache.pinot.query.planner.nodes.RexExpression;
 import org.apache.pinot.segment.spi.AggregationFunctionType;
 import org.apache.pinot.sql.parsers.SqlCompilationException;
 import org.slf4j.Logger;
@@ -50,10 +45,10 @@ import org.slf4j.LoggerFactory;
 
 
 /**
- * Calcite parser to convert SQL expressions into {@link Expression}.
+ * This class provide API to parse a SQL string into Pinot query {@link SqlNode}.
  *
- * <p>This class is extracted from {@link org.apache.pinot.sql.parsers.CalciteSqlParser}. It only contains the
- * {@link Expression} related info, this is used for ingestion and query rewrite.
+ * <p>This class is extracted from {@link org.apache.pinot.sql.parsers.CalciteSqlParser}. It contains the logic
+ * to parsed SQL in the format of {@link SqlNode} and convert them into Thrift {@link Expression} format.
  */
 public class CalciteExpressionParser {
   private static final Logger LOGGER = LoggerFactory.getLogger(CalciteExpressionParser.class);
@@ -351,7 +346,7 @@ public class CalciteExpressionParser {
         operands.add(toExpression(childNode));
       }
     }
-    validateFunction(functionName, operands);
+    ParserUtils.validateFunction(functionName, operands);
     Expression functionExpression = RequestUtils.getFunctionExpression(functionName);
     functionExpression.getFunctionCall().setOperands(operands);
     return functionExpression;
@@ -407,57 +402,6 @@ public class CalciteExpressionParser {
     }
   }
 
-  public static String canonicalize(String functionName) {
-    return StringUtils.remove(functionName, '_').toLowerCase();
-  }
-
-  public static boolean isSameFunction(String function1, String function2) {
-    return canonicalize(function1).equals(canonicalize(function2));
-  }
-
-  private static void validateFunction(String functionName, List<Expression> operands) {
-    switch (canonicalize(functionName)) {
-      case "jsonextractscalar":
-        validateJsonExtractScalarFunction(operands);
-        break;
-      case "jsonextractkey":
-        validateJsonExtractKeyFunction(operands);
-        break;
-      default:
-        break;
-    }
-  }
-
-  private static void validateJsonExtractScalarFunction(List<Expression> operands) {
-    int numOperands = operands.size();
-
-    // Check that there are exactly 3 or 4 arguments
-    if (numOperands != 3 && numOperands != 4) {
-      throw new SqlCompilationException(
-          "Expect 3 or 4 arguments for transform function: jsonExtractScalar(jsonFieldName, 'jsonPath', "
-              + "'resultsType', ['defaultValue'])");
-    }
-    if (!operands.get(1).isSetLiteral() || !operands.get(2).isSetLiteral() || (numOperands == 4 && !operands.get(3)
-        .isSetLiteral())) {
-      throw new SqlCompilationException(
-          "Expect the 2nd/3rd/4th argument of transform function: jsonExtractScalar(jsonFieldName, 'jsonPath',"
-              + " 'resultsType', ['defaultValue']) to be a single-quoted literal value.");
-    }
-  }
-
-  private static void validateJsonExtractKeyFunction(List<Expression> operands) {
-    // Check that there are exactly 2 arguments
-    if (operands.size() != 2) {
-      throw new SqlCompilationException(
-          "Expect 2 arguments are required for transform function: jsonExtractKey(jsonFieldName, 'jsonPath')");
-    }
-    if (!operands.get(1).isSetLiteral()) {
-      throw new SqlCompilationException(
-          "Expect the 2nd argument for transform function: jsonExtractKey(jsonFieldName, 'jsonPath') to be a "
-              + "single-quoted literal value.");
-    }
-  }
-
   /**
    * Helper method to flatten the operands for the AND expression.
    */
@@ -506,170 +450,5 @@ public class CalciteExpressionParser {
       return false;
     }
     return false;
-  }
-
-  // --------------------------------------------------------------------------
-  // Relational conversion Utils
-  // --------------------------------------------------------------------------
-
-  private static List<Expression> convertDistinctSelectList(RexExpression.FunctionCall rexCall, PinotQuery pinotQuery) {
-    List<Expression> selectExpr = new ArrayList<>();
-    selectExpr.add(convertDistinctAndSelectListToFunctionExpression(rexCall, pinotQuery));
-    return selectExpr;
-  }
-
-  public static List<Expression> convertSelectList(List<RexExpression> rexNodeList, PinotQuery pinotQuery) {
-    List<Expression> selectExpr = new ArrayList<>();
-
-    final Iterator<RexExpression> iterator = rexNodeList.iterator();
-    while (iterator.hasNext()) {
-      final RexExpression next = iterator.next();
-      selectExpr.add(toExpression(next, pinotQuery));
-    }
-
-    return selectExpr;
-  }
-
-  private static List<Expression> convertOrderByList(RexExpression.FunctionCall rexCall, PinotQuery pinotQuery) {
-    Preconditions.checkState(rexCall.getKind() == SqlKind.ORDER_BY);
-    List<Expression> orderByExpr = new ArrayList<>();
-
-    final Iterator<RexExpression> iterator = rexCall.getFunctionOperands().iterator();
-    while (iterator.hasNext()) {
-      final RexExpression next = iterator.next();
-      orderByExpr.add(convertOrderBy(next, pinotQuery));
-    }
-    return orderByExpr;
-  }
-
-  private static Expression convertOrderBy(RexExpression rexNode, PinotQuery pinotQuery) {
-    final SqlKind kind = rexNode.getKind();
-    Expression expression;
-    switch (kind) {
-      case DESCENDING:
-        RexExpression.FunctionCall rexCall = (RexExpression.FunctionCall) rexNode;
-        expression = RequestUtils.getFunctionExpression("DESC");
-        expression.getFunctionCall().addToOperands(toExpression(rexCall.getFunctionOperands().get(0), pinotQuery));
-        break;
-      case IDENTIFIER:
-      default:
-        expression = RequestUtils.getFunctionExpression("ASC");
-        expression.getFunctionCall().addToOperands(toExpression(rexNode, pinotQuery));
-        break;
-    }
-    return expression;
-  }
-
-  private static Expression convertDistinctAndSelectListToFunctionExpression(RexExpression.FunctionCall rexCall,
-      PinotQuery pinotQuery) {
-    String functionName = AggregationFunctionType.DISTINCT.getName();
-    Expression functionExpression = RequestUtils.getFunctionExpression(functionName);
-    for (RexExpression node : rexCall.getFunctionOperands()) {
-      Expression columnExpression = toExpression(node, pinotQuery);
-      if (columnExpression.getType() == ExpressionType.IDENTIFIER && columnExpression.getIdentifier().getName()
-          .equals("*")) {
-        throw new SqlCompilationException(
-            "Syntax error: Pinot currently does not support DISTINCT with *. Please specify each column name after "
-                + "DISTINCT keyword");
-      } else if (columnExpression.getType() == ExpressionType.FUNCTION) {
-        Function functionCall = columnExpression.getFunctionCall();
-        String function = functionCall.getOperator();
-        if (AggregationFunctionType.isAggregationFunction(function)) {
-          throw new SqlCompilationException(
-              "Syntax error: Use of DISTINCT with aggregation functions is not supported");
-        }
-      }
-      functionExpression.getFunctionCall().addToOperands(columnExpression);
-    }
-    return functionExpression;
-  }
-
-  public static Expression toExpression(RexExpression rexNode, PinotQuery pinotQuery) {
-    LOGGER.debug("Current processing RexNode: {}, node.getKind(): {}", rexNode, rexNode.getKind());
-    switch (rexNode.getKind()) {
-      case INPUT_REF:
-        return inputRefToIdentifier((RexExpression.InputRef) rexNode, pinotQuery);
-      case LITERAL:
-        return rexLiteralToExpression((RexExpression.Literal) rexNode);
-      default:
-        return compileFunctionExpression((RexExpression.FunctionCall) rexNode, pinotQuery);
-    }
-  }
-
-  private static Expression rexLiteralToExpression(RexExpression.Literal rexLiteral) {
-    RelDataType type = rexLiteral.getDataType();
-    switch (type.getSqlTypeName()) {
-      default:
-        return RequestUtils.getLiteralExpression(rexLiteral.getValue());
-    }
-  }
-
-  public static Expression inputRefToIdentifier(RexExpression.InputRef inputRef, PinotQuery pinotQuery) {
-    List<Expression> selectList = pinotQuery.getSelectList();
-    return selectList.get(inputRef.getIndex());
-  }
-
-  private static Expression compileFunctionExpression(RexExpression.FunctionCall rexCall, PinotQuery pinotQuery) {
-    SqlKind functionKind = rexCall.getKind();
-    String functionName;
-    switch (functionKind) {
-      case AND:
-        return compileAndExpression(rexCall, pinotQuery);
-      case OR:
-        return compileOrExpression(rexCall, pinotQuery);
-      case COUNT:
-      case OTHER:
-      case OTHER_FUNCTION:
-      case DOT:
-      default:
-        functionName = functionKind.name();
-        break;
-    }
-    // When there is no argument, set an empty list as the operands
-    List<RexExpression> childNodes = rexCall.getFunctionOperands();
-    List<Expression> operands = new ArrayList<>(childNodes.size());
-    for (RexExpression childNode : childNodes) {
-      operands.add(toExpression(childNode, pinotQuery));
-    }
-    validateFunction(functionName, operands);
-    Expression functionExpression = RequestUtils.getFunctionExpression(functionName);
-    functionExpression.getFunctionCall().setOperands(operands);
-    return functionExpression;
-  }
-
-  /**
-   * Helper method to flatten the operands for the AND expression.
-   */
-  private static Expression compileAndExpression(RexExpression.FunctionCall andNode, PinotQuery pinotQuery) {
-    List<Expression> operands = new ArrayList<>();
-    for (RexExpression childNode : andNode.getFunctionOperands()) {
-      if (childNode.getKind() == SqlKind.AND) {
-        Expression childAndExpression = compileAndExpression((RexExpression.FunctionCall) childNode, pinotQuery);
-        operands.addAll(childAndExpression.getFunctionCall().getOperands());
-      } else {
-        operands.add(toExpression(childNode, pinotQuery));
-      }
-    }
-    Expression andExpression = RequestUtils.getFunctionExpression(SqlKind.AND.name());
-    andExpression.getFunctionCall().setOperands(operands);
-    return andExpression;
-  }
-
-  /**
-   * Helper method to flatten the operands for the OR expression.
-   */
-  private static Expression compileOrExpression(RexExpression.FunctionCall orNode, PinotQuery pinotQuery) {
-    List<Expression> operands = new ArrayList<>();
-    for (RexExpression childNode : orNode.getFunctionOperands()) {
-      if (childNode.getKind() == SqlKind.OR) {
-        Expression childAndExpression = compileOrExpression((RexExpression.FunctionCall) childNode, pinotQuery);
-        operands.addAll(childAndExpression.getFunctionCall().getOperands());
-      } else {
-        operands.add(toExpression(childNode, pinotQuery));
-      }
-    }
-    Expression andExpression = RequestUtils.getFunctionExpression(SqlKind.OR.name());
-    andExpression.getFunctionCall().setOperands(operands);
-    return andExpression;
   }
 }
