+++ pinot-query-runtime/src/test/java/org/apache/pinot/query/service/QueryServerTest.java
@@ -132,7 +132,7 @@ public class QueryServerTest {
         new Object[]{"SELECT * FROM a"},
         new Object[]{"SELECT * FROM a JOIN b ON a.col3 = b.col3"},
         new Object[]{"SELECT a.col1, a.ts, c.col2, c.col3 FROM a JOIN c ON a.col1 = c.col2 "
-            + " WHERE a.col3 >= 0 AND a.col2 = 'foo' AND c.col3 >= 0"},
+            + " WHERE (a.col3 >= 0 OR a.col2 = 'foo') AND c.col3 >= 0"},
     };
   }
